+++ apps/web/examples/dashboard/components/user-nav.tsx
@@ -17,7 +17,7 @@ export function UserNav() {
       <DropdownMenuTrigger asChild>
         <Button variant="ghost" className="relative h-8 w-8 rounded-full">
           <Avatar className="h-8 w-8">
-            <AvatarImage src="/avatars/01.png" alt="@dinsterizer" />
+            <AvatarImage src="/avatars/01.png" alt="@dinwwwh" />
             <AvatarFallback>DN</AvatarFallback>
           </Avatar>
         </Button>
@@ -25,8 +25,8 @@ export function UserNav() {
       <DropdownMenuContent className="w-56" align="end" forceMount>
         <DropdownMenuLabel className="font-normal">
           <div className="flex flex-col space-y-1">
-            <p className="text-sm font-medium leading-none">Din</p>
-            <p className="text-xs leading-none text-muted-foreground">dinsterizer@gmail.com</p>
+            <p className="text-sm font-medium leading-none">dinwwwh</p>
+            <p className="text-xs leading-none text-muted-foreground">dinwwwh@gmail.com</p>
           </div>
         </DropdownMenuLabel>
         <DropdownMenuSeparator />