+++ apps/web/examples/ui/input-form.tsx
@@ -51,7 +51,7 @@ export default function InputForm() {
             <FormItem>
               <FormLabel>Username</FormLabel>
               <FormControl>
-                <Input placeholder="dinsterizer" {...field} />
+                <Input placeholder="dinwwwh" {...field} />
               </FormControl>
               <FormDescription>This is your public display name.</FormDescription>
               <FormMessage />