+++ packages/cli/env.ts
@@ -1,5 +1,5 @@
 export const env = {
   PROJECT_NAME: 'DinUI',
-  PROBLEM_URL: 'https://github.com/dinsterizer/dinui/issues',
-  DOCS_URL: 'https://ui.dinsterizer.com/introduction',
+  PROBLEM_URL: 'https://github.com/dinwwwh/dinui/issues',
+  DOCS_URL: 'https://ui.dinwwwh.com/introduction',
 }