+++ packages/cli/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Minor Changes
 
-- [#19](https://github.com/dinsterizer/dinui/pull/19) [`17cfe44`](https://github.com/dinsterizer/dinui/commit/17cfe441a3d33cceea89e24f318ca78340d7ffee) Thanks [@dinsterizer](https://github.com/dinsterizer)! - new `add` command for adding one or many components to your project
+- [#19](https://github.com/dinwwwh/dinui/pull/19) [`17cfe44`](https://github.com/dinwwwh/dinui/commit/17cfe441a3d33cceea89e24f318ca78340d7ffee) Thanks [@dinwwwh](https://github.com/dinwwwh)! - new `add` command for adding one or many components to your project